requirements.txt
# Core Anomalib and PyTorch dependencies
anomalib>=1.0.0
torch>=1.13.0
torchvision>=0.14.0
pytorch-lightning>=1.9.0

# Data science and machine learning
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.3.0
imageio>=2.9.0

# Configuration and utilities
omegaconf>=2.1.0
hydra-core>=1.1.0
PyYAML>=6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Logging and monitoring (optional but recommended)
tensorboard>=2.8.0
wandb>=0.12.0

# Development and testing
pytest>=6.2.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.910

# Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=7.6.0

# Additional utilities
tqdm>=4.62.0
rich>=10.0.0

# OpenVINO (for model optimization - optional)
# openvino-dev>=2022.1.0

# GPU support (uncomment if using CUDA)
# torch-audio  # for additional audio processing if needed